responses

+++ src/chat/chat.gateway.ts
@@ -8,6 +8,7 @@ import {
 } from '@nestjs/websockets';
 import { Server, Socket } from 'socket.io';
 import { ChatService } from './chat.service';
+import { MessageDto } from './schemas/chat.schema';
 
 @WebSocketGateway(3002, {
   cors: {
@@ -25,16 +26,20 @@ export class ChatGateway implements OnGatewayInit {
 
   @SubscribeMessage('sendMessage')
   async handleMessage(
-    @MessageBody() data: any,
+    @MessageBody() data: MessageDto,
   ) {
     const savedMessage = await this.chatService.saveMessage(data);
     this.server.to(data.receiverId).emit('receiveMessage', savedMessage);
     this.server.to(data.senderId).emit('receiveMessage', savedMessage);
+
+    return savedMessage;
   }
 
   @SubscribeMessage('joinRoom')
   handleJoinRoom(@MessageBody() userId: string, @ConnectedSocket() client: Socket) {
-    client.join(userId); // each user joins their own room
+    if (!userId) return;
+
+    client.join(userId);
   }
 
   @SubscribeMessage('deleteMessage')
@@ -42,6 +47,12 @@ export class ChatGateway implements OnGatewayInit {
     @MessageBody() data: { messageId: string; senderId: string; receiverId: string }
   ) {
     const { messageId, senderId, receiverId } = data;
+
+    console.log(data);
+    
+
+    if (!messageId || !senderId || !receiverId) return;
+
     await this.chatService.deleteMessage(messageId);
 
     this.server.to(senderId).emit('messageDeleted', { messageId, success: true });
@@ -51,22 +62,39 @@ export class ChatGateway implements OnGatewayInit {
   @SubscribeMessage('acceptMessage')
   async handleAcceptMessage(@MessageBody() data: any) {
     const { messageId, senderId, receiverId, name, orderId } = data;
+
+    if (!messageId || !senderId || !receiverId || !orderId) return;
+
     await this.chatService.acceptMessage(messageId, orderId, receiverId);
 
     this.server.to(senderId).emit('messageAccepted', { messageId, success: true });
     this.server.to(receiverId).emit('messageAccepted', { messageId, success: true });
 
-    this.handleMessage({ senderId, receiverId, text: `Предложение было принято ${name}.`, status: 'server' });
+    this.handleMessage({ senderId, receiverId, text: `Предложение было принято${name ? (' ' + name) : ''}.`, status: 'server', is_suggest: false, mediaType: 'none', mediaUrl: '', createdAt: '' });
   }
 
   @SubscribeMessage('rejectMessage')
   async handleRejectMessage(@MessageBody() data: any) {
     const { messageId, senderId, receiverId, name } = data;
+
+    if (!messageId || !senderId || !receiverId) return;
+
     await this.chatService.rejectMessage(messageId);
 
     this.server.to(senderId).emit('messageRejected', { messageId, success: true });
     this.server.to(receiverId).emit('messageRejected', { messageId, success: true });
 
-    this.handleMessage({ senderId, receiverId, text: `Предложение было отклонено ${name}.`, status: 'server' });
+    this.handleMessage({ senderId, receiverId, text: `Предложение было отклонено${name ? (' ' + name) : ''}.`, status: 'server', is_suggest: false, mediaType: 'none', mediaUrl: '', createdAt: '' });
+  }
+
+  @SubscribeMessage('postResponse')
+  async handlePostResponse(@MessageBody() data: any) {
+    const { senderId, receiverId, orderId, responseId } = data;
+
+    if (!senderId || !receiverId || !orderId || !responseId) return;
+
+    const res = await this.handleMessage({ senderId, receiverId, text: `Отклик на заказ`, status: 'response', is_suggest: false, mediaType: 'none', mediaUrl: '', createdAt: '', orderId, responseId });
+
+    this.chatService.patchResponse(responseId, res.id);
   }
 }
